Insertion working Code refactor / Folder structure Adding traits

+++ src/main/scala/be/serverunit/api/JsonExtractor.scala
@@ -6,7 +6,6 @@ import java.time.LocalDateTime
 
 object JsonExtractor {
   def extractStartData(json: JsValue): Option[(String, LocalDateTime, Int)] = {
-    // Extracting the data from the json (user, timer) using a for comprehension
     for {
       user <- (json \ "user").asOpt[String]
       time <- (json \ "time").asOpt[LocalDateTime]
@@ -14,16 +13,14 @@ object JsonExtractor {
     } yield (user, time, weight)
   }
 
-  def extractData(json: JsValue): Option[(Int, Int)] = {
-    // Extracting the data from the json (user, distance, timer) using a for comprehension
+  def extractData(json: JsValue): Option[(Int, Float)] = {
     for {
       distance <- (json \ "distance").asOpt[Int]
-      timer <- (json \ "timer").asOpt[Int]
+      timer <- (json \ "timer").asOpt[Float]
     } yield (distance, timer)
   }
 
   def extractEndData(json: JsValue): Option[(Int, LocalDateTime)] = {
-    // Extracting the data from the json (user, reps, time) using a for comprehension
     for {
       user <- (json \ "user").asOpt[String]
       reps <- (json \ "reps").asOpt[Int]
@@ -32,7 +29,6 @@ object JsonExtractor {
   }
 
   def extractAirData(json: JsValue): Option[(Float, Float, Float, LocalDateTime)] = {
-    // Extracting the data from the json (temperature, humidity, pm, timestamp) using a for comprehension
     for {
       temperature <- (json \ "temperature").asOpt[Float]
       humidity <- (json \ "humidity").asOpt[Float]
